backend/requirements.txt
uvicorn
python-multipart
torch
torchvision
torchaudio
numpy
pandas
scikit-learn
librosa
soundfile
pillow
joblib
requests
fastapi